+++ packages/terra-clinical-detail-view/tests/wdio/DetailView/detail-view-spec.js
@@ -11,17 +11,7 @@ Terra.viewports('tiny', 'small', 'medium', 'large', 'huge').forEach((viewport) =
         browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided');
       });
 
-      Terra.should.validateElement({ selector });
-      Terra.should.themeCombinationOfCustomProperties({
-        testName: 'themed',
-        selector,
-        properties: {
-          '--terra-clinical-detail-view-subtitle-color': 'purple',
-          '--terra-clinical-detail-view-accessory-color': 'red',
-          '--terra-clinical-detail-view-divider-border-color': 'yellow',
-          '--terra-clinical-detail-view-footer-text-color': 'green',
-        },
-      });
+      Terra.it.validatesElement({ selector });
     });
 
     describe('Displays a Detail View that is not divided with provided components', () => {
@@ -29,7 +19,7 @@ Terra.viewports('tiny', 'small', 'medium', 'large', 'huge').forEach((viewport) =
         browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-no-divider');
       });
 
-      Terra.should.validateElement({ selector });
+      Terra.it.validatesElement({ selector });
     });
 
     describe('Displays a Detail View with provided components and when title size is smaller', () => {
@@ -37,13 +27,13 @@ Terra.viewports('tiny', 'small', 'medium', 'large', 'huge').forEach((viewport) =
         browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-divided-smaller-titles');
       });
 
-      Terra.should.validateElement({ selector });
+      Terra.it.validatesElement({ selector });
     });
 
     describe('Displays a Detail View with overflowed title text', () => {
       before(() => browser.url('/#/raw/tests/terra-clinical-detail-view/clinical-detail-view/detail-view/detail-view-overflow-text'));
 
-      Terra.should.validateElement({ selector });
+      Terra.it.validatesElement({ selector });
     });
   });
 });