+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -8,6 +8,9 @@ Unreleased
 * Indicate in site examples what birthdate is set to.
 * Added legend and isLegendHidden prop to Onsetpicker field group.
 
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 4.2.0 - (April 16, 2019)
 ----------
 ### Fixed